Introduce cw_text_parser.recalculate_speeds()
This way, the delays for dits, dahs etc. are stored as class variables.
So far, a new word was keyed at the default speed. With this change,
the keying speed can be persistent between words, i.e. speeding up with
<+> holds until the <-> macro is used, even if the current word ends
and a new word starts. This does not work yet, though, and will be
implemented when self.wpm is introduced (to track WPM between words).